Day11. Quiz1 Solved. Refactored from int values to duble so bigger worry levels can be handled

+++ src/day11/Operation.groovy
@@ -10,7 +10,7 @@ class Operation {
 	
 	String operator
 	
-	def OPERANDS = ["+":{int l, int r -> l+r}, "*":{int l, int r -> l*r}]
+	def OPERANDS = ["+":{double l, double r -> l+r}, "*":{double l, double r -> l*r}]
 	
 	
 	Operation(def leftOperand, String operator, def rightOperand) {
@@ -24,14 +24,14 @@ class Operation {
 	}
 	
 	
-	int getLeftOperandValue(int worryLevel) {
+	double getLeftOperandValue(double worryLevel) {
 		if (leftOperand == OLD) {
 			return worryLevel
 		}
 		return leftOperand as Integer
 	}
 	
-	int getRightOperandValue(int worryLevel) {
+	double getRightOperandValue(double worryLevel) {
 		if (rightOperand == OLD) {
 			return worryLevel
 		}
@@ -39,10 +39,10 @@ class Operation {
 		return rightOperand as Integer
 	}
 	
-	int execute(int worryLevel) {
+	double execute(double worryLevel) {
 		
-		int leftValue = getLeftOperandValue(worryLevel)
-		int rightValue = getRightOperandValue(worryLevel)
+		double leftValue = getLeftOperandValue(worryLevel)
+		double rightValue = getRightOperandValue(worryLevel)
 		
 		return OPERANDS[operator].call(leftValue, rightValue)
 	}
